# Re-write JsonParser on top of the JsonReader APIs