--- a/samples/LowAllocationWebServer/Shared/SampleServer.cs
+++ b/samples/LowAllocationWebServer/Shared/SampleServer.cs
@@ -80,7 +80,8 @@ namespace LowAllocationWebServer
             int requestedCount;
 
             // TODO: this should not convert to span
-            var dom = JsonObject.Parse(body.First.Span);
+            var parser = new JsonParser(body.First.Span);
+            JsonObject dom = parser.Parse();
             try
             {
                 requestedCount = (int)dom["Count"];