--- a/tests/Benchmarks/System.Text.JsonLab/JsonParserPerf.cs
+++ b/tests/Benchmarks/System.Text.JsonLab/JsonParserPerf.cs
@@ -45,7 +45,7 @@ namespace System.Text.JsonLab.Benchmarks
         }
 
         [Benchmark]
-        public void ReaderSystemTextJsonLab()
+        public void ParseSystemTextJsonLab()
         {
             JsonObject.Parse(_dataUtf8);
         }