Using the JsonReader APIs to build a JsonParser on top, rather than using custom parsing logic (required adding Index and StartLocation properties to JsonReader).

Also changed the stack representation (implementation detail) and consolidated duplicate enums.

**Before:** https://github.com/dotnet/corefxlab/pull/2431#issue-208456442

**After:** 50-300% performance improvement (although number parsing/validation is missing)

``` ini

BenchmarkDotNet=v0.10.14.683-nightly, OS=Windows 10.0.17713
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-20180720-2
  [Host]     : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  Job-JSHUMY : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT

IterationCount=5  WarmupCount=3  

```
|                  Method |        TestCase |           Mean |         Error |        StdDev |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------------------------ |---------------- |---------------:|--------------:|--------------:|---------:|---------:|---------:|----------:|
| **ReaderSystemTextJsonLab** |       **BasicJson** |     **1,189.7 ns** |     **213.64 ns** |      **55.49 ns** |   **0.0038** |        **-** |        **-** |      **24 B** |
| **ReaderSystemTextJsonLab** |   **BasicLargeNum** |     **1,430.5 ns** |      **66.52 ns** |      **17.28 ns** |   **0.0038** |        **-** |        **-** |      **24 B** |
| **ReaderSystemTextJsonLab** |       **BroadTree** |    **28,711.7 ns** |   **6,258.17 ns** |   **1,625.54 ns** |        **-** |        **-** |        **-** |      **24 B** |
| **ReaderSystemTextJsonLab** |        **DeepTree** |    **13,165.0 ns** |     **428.36 ns** |     **111.26 ns** |        **-** |        **-** |        **-** |      **24 B** |
| **ReaderSystemTextJsonLab** |     **FullSchema1** |     **3,557.0 ns** |     **804.73 ns** |     **209.03 ns** |   **0.0038** |        **-** |        **-** |      **24 B** |
| **ReaderSystemTextJsonLab** |     **FullSchema2** |     **6,126.6 ns** |     **493.18 ns** |     **128.10 ns** |        **-** |        **-** |        **-** |      **24 B** |
| **ReaderSystemTextJsonLab** |      **HelloWorld** |       **341.6 ns** |     **146.82 ns** |      **38.14 ns** |   **0.0052** |        **-** |        **-** |      **24 B** |
| **ReaderSystemTextJsonLab** |        **Json400B** |     **1,403.0 ns** |     **113.97 ns** |      **29.60 ns** |   **0.0038** |        **-** |        **-** |      **24 B** |
| **ReaderSystemTextJsonLab** |       **Json400KB** | **2,754,621.2 ns** | **358,422.89 ns** |  **93,099.01 ns** | **332.0313** | **332.0313** | **332.0313** | **1711944 B** |
| **ReaderSystemTextJsonLab** |        **Json40KB** |   **145,131.1 ns** |   **2,100.96 ns** |     **545.72 ns** |        **-** |        **-** |        **-** |      **24 B** |
| **ReaderSystemTextJsonLab** |         **Json4KB** |    **13,299.5 ns** |   **3,082.93 ns** |     **800.78 ns** |        **-** |        **-** |        **-** |      **24 B** |
| **ReaderSystemTextJsonLab** |   **LotsOfNumbers** |     **5,804.4 ns** |   **2,750.89 ns** |     **714.53 ns** |        **-** |        **-** |        **-** |      **24 B** |
| **ReaderSystemTextJsonLab** |   **LotsOfStrings** |     **3,136.5 ns** |   **1,431.04 ns** |     **371.71 ns** |   **0.0038** |        **-** |        **-** |      **24 B** |
| **ReaderSystemTextJsonLab** | **ProjectLockJson** |   **928,512.6 ns** | **408,032.81 ns** | **105,985.00 ns** |        **-** |        **-** |        **-** |      **24 B** |
| **ReaderSystemTextJsonLab** |  **SpecialNumForm** |     **2,212.8 ns** |     **335.88 ns** |      **87.24 ns** |   **0.0038** |        **-** |        **-** |      **24 B** |