@ahsonkhan 2018-08-17
Additional 20-40% performance improvement (+ removing the allocations).
- Using ArrayPool instead.
- Reducing the size of allocated buffer.
- Traverse database from the bottom for quicker search.

``` ini

BenchmarkDotNet=v0.10.14.683-nightly, OS=Windows 10.0.17713
Intel Core i7-6700 CPU 3.40GHz (Max: 2.90GHz) (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-20180720-2
  [Host]     : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  Job-GYIZAE : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT

IterationCount=5  WarmupCount=3  

```
|                 Method |        TestCase |           Mean |          Error |        StdDev | Allocated |
|----------------------- |---------------- |---------------:|---------------:|--------------:|----------:|
| **ParseSystemTextJsonLab** |       **BasicJson** |     **1,005.1 ns** |      **33.588 ns** |      **8.724 ns** |       **0 B** |
| **ParseSystemTextJsonLab** |   **BasicLargeNum** |     **1,200.8 ns** |      **44.301 ns** |     **11.507 ns** |       **0 B** |
| **ParseSystemTextJsonLab** |       **BroadTree** |    **17,287.2 ns** |     **533.138 ns** |    **138.481 ns** |       **0 B** |
| **ParseSystemTextJsonLab** |        **DeepTree** |    **11,533.3 ns** |     **230.527 ns** |     **59.879 ns** |       **0 B** |
| **ParseSystemTextJsonLab** |     **FullSchema1** |     **2,703.1 ns** |      **43.377 ns** |     **11.267 ns** |       **0 B** |
| **ParseSystemTextJsonLab** |     **FullSchema2** |     **3,863.3 ns** |     **130.306 ns** |     **33.846 ns** |       **0 B** |
| **ParseSystemTextJsonLab** |      **HelloWorld** |       **250.6 ns** |       **9.577 ns** |      **2.487 ns** |       **0 B** |
| **ParseSystemTextJsonLab** |        **Json400B** |     **1,360.8 ns** |      **21.345 ns** |      **5.544 ns** |       **0 B** |
| **ParseSystemTextJsonLab** |       **Json400KB** | **1,201,668.2 ns** | **218,919.624 ns** | **56,863.558 ns** |       **0 B** |
| **ParseSystemTextJsonLab** |        **Json40KB** |   **113,309.4 ns** |   **1,762.016 ns** |    **457.677 ns** |       **0 B** |
| **ParseSystemTextJsonLab** |         **Json4KB** |    **10,482.4 ns** |     **444.018 ns** |    **115.332 ns** |       **0 B** |
| **ParseSystemTextJsonLab** |   **LotsOfNumbers** |     **5,317.3 ns** |     **946.814 ns** |    **245.931 ns** |       **0 B** |
| **ParseSystemTextJsonLab** |   **LotsOfStrings** |     **2,625.6 ns** |     **204.172 ns** |     **53.033 ns** |       **0 B** |
| **ParseSystemTextJsonLab** | **ProjectLockJson** |   **456,216.0 ns** |   **4,402.400 ns** |  **1,143.507 ns** |       **0 B** |
| **ParseSystemTextJsonLab** |  **SpecialNumForm** |     **1,990.5 ns** |      **90.843 ns** |     **23.596 ns** |       **0 B** |